requirements.txt
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.7
idna==2.10
phonenumbers==8.12.7
pypiwin32==223
pyttsx3==2.90
pywin32==228
requests==2.24.0
SpeechRecognition==3.8.1
speedtest-cli==2.1.2
urllib3==1.25.10
wincertstore==0.2
